backend/requirements.txt
Django==6.0.4
djangorestframework==3.17.1
django-cors-headers==4.9.0
gunicorn==25.3.0
psycopg2-binary==2.9.12
python-dotenv==1.2.2
requests==2.33.1
pypdf==6.10.2
python-docx==1.2.0
django-ratelimit==4.1.0
